refactor(sasjs-command-tests): refactored file structure
